+++ server/lib/config/index.js
@@ -2,6 +2,7 @@ const _ = require('lodash');
 const appLog = require('../app-log');
 const configItems = require('./config-items');
 const validateConnection = require('../validate-connection');
+const removedEnv = require('./removed-env');
 const fromDefault = require('./from-default');
 const fromEnv = require('./from-env');
 const fromCli = require('./from-cli');
@@ -11,6 +12,7 @@ const getOldConfigWarning = require('./get-old-config-warning');
 class Config {
   constructor(argv, env) {
     this.argv = argv;
+    this.env = env;
 
     const configFilePath = argv.config || env.SQLPAD_CONFIG;
 
@@ -62,10 +64,41 @@ class Config {
       errors.push(getOldConfigWarning());
     }
 
-    // Check for any unknown or deprecated keys provided in config
-    // Connections key is filtered out from consideration here because it is special and dynamic
-    // When dealing with unknown keys, the unknown key will only ever come from a file or CLI.
-    // Config from environment vars will only ever be plucked from the known list of config items
+    // Check for any old environment variables in env.
+    // This must be handled separately from other unknown checks,
+    // as fromEnv() only gets config it knows about, so it will never have unknown values
+    removedEnv.forEach(key => {
+      if (this.env.hasOwnProperty(key)) {
+        errors.push(
+          `CONFIG NOT RECOGNIZED: Environment variable "${key}" no longer supported.`
+        );
+      }
+    });
+
+    // Check CLI config for any unknown flags
+    // Flag must be from config-items + select values (-v -h --version --help, _ is used for all non-named values)
+    const additionalCliFlags = ['_', 'h', 'help', 'v', 'version'];
+    Object.keys(this.argv).forEach(key => {
+      const inAdditional = additionalCliFlags.includes(key);
+      const inConfigItems = Boolean(configItems.find(item => item.key === key));
+      if (!inAdditional && !inConfigItems) {
+        errors.push(`CONFIG NOT RECOGNIZED: cli flag "${key}"`);
+      }
+    });
+
+    // Check fileConfig for unknown keys
+    // Connections key is filtered out from consideration here because it is not driven by config-items
+    // If key is not found in config items, raise error message to prevent application startup
+    Object.keys(this.fileConfig)
+      .filter(key => key !== 'connections')
+      .filter(key => !configItems.find(item => item.key === key))
+      .forEach(key => {
+        errors.push(
+          `CONFIG NOT RECOGNIZED: Key "${key}" in file ${this.configFilePath}.`
+        );
+      });
+
+    // Check for deprecated keys provided in config
     const userProvidedConfigs = {
       ...this.envConfig,
       ...this.fileConfig,
@@ -75,19 +108,7 @@ class Config {
       .filter(key => key !== 'connections')
       .forEach(key => {
         const configItem = configItems.find(item => item.key === key);
-        if (!configItem) {
-          if (this.fileConfig.hasOwnProperty(key)) {
-            warnings.push(
-              `CONFIG NOT RECOGNIZED: Key ${key} in file ${this.configFilePath}.`
-            );
-          }
-          if (this.cliConfig.hasOwnProperty(key)) {
-            warnings.push(`CONFIG NOT RECOGNIZED: CLI flag ${key}.`);
-          }
-          return;
-        }
-
-        if (configItem.deprecated) {
+        if (configItem && configItem.deprecated) {
           warnings.push(
             `DEPRECATED CONFIG: ${configItem.key} / ${configItem.envVar}. ${configItem.deprecated}`
           );